projects/vision_aiii/requirements.txt
torch
torchvision
opencv-python
numpy
scipy
Pillow
matplotlib
ultralytics
